+++ pkg/compactor/shard_aware_deduplicate_filter_test.go
@@ -391,7 +391,6 @@ func newTestFetcherMetrics() *block.FetcherMetrics {
 }
 
 func BenchmarkDeduplicateFilter_Filter(b *testing.B) {
-
 	var (
 		reg   prometheus.Registerer
 		count uint64