+++ pkg/distributor/push.go
@@ -226,6 +226,8 @@ func toHTTPStatus(ctx context.Context, pushErr error, limits *validation.Overrid
 			return http.StatusBadRequest
 		case mimirpb.TSDB_UNAVAILABLE:
 			return http.StatusServiceUnavailable
+		case mimirpb.CIRCUIT_BREAKER_OPEN:
+			return http.StatusServiceUnavailable
 		}
 	}