+++ pkg/ruler/rulestore/config.go
@@ -15,8 +15,11 @@ import (
 	"github.com/grafana/dskit/cache"
 	"github.com/grafana/dskit/flagext"
 
-	"github.com/grafana/mimir/pkg/ruler/rulestore/local"
 	"github.com/grafana/mimir/pkg/storage/bucket"
+)
+
+const (
+	BackendLocal = "local"
 )
 
 var supportedCacheBackends = []string{cache.BackendMemcached, cache.BackendRedis}
@@ -24,7 +27,7 @@ var supportedCacheBackends = []string{cache.BackendMemcached, cache.BackendRedis
 // Config configures a rule store.
 type Config struct {
 	bucket.Config `yaml:",inline"`
-	Local         local.Config `yaml:"local"`
+	Local         LocalStoreConfig `yaml:"local"`
 
 	// Cache holds the configuration used for the ruler storage cache.
 	Cache cache.BackendConfig `yaml:"cache"`
@@ -34,11 +37,11 @@ type Config struct {
 func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	prefix := "ruler-storage."
 
-	cfg.StorageBackendConfig.ExtraBackends = []string{local.Name}
+	cfg.StorageBackendConfig.ExtraBackends = []string{BackendLocal}
 	cfg.Local.RegisterFlagsWithPrefix(prefix, f)
 	cfg.RegisterFlagsWithPrefixAndDefaultDirectory(prefix, "ruler", f)
 
-	f.StringVar(&cfg.Cache.Backend, prefix+"cache.backend", "", fmt.Sprintf("Backend for ruler storage cache, if not empty. The cache is supported for any storage backend except %q. Supported values: %s.", local.Name, strings.Join(supportedCacheBackends, ", ")))
+	f.StringVar(&cfg.Cache.Backend, prefix+"cache.backend", "", fmt.Sprintf("Backend for ruler storage cache, if not empty. The cache is supported for any storage backend except %q. Supported values: %s.", BackendLocal, strings.Join(supportedCacheBackends, ", ")))
 	cfg.Cache.Memcached.RegisterFlagsWithPrefix(prefix+"cache.memcached.", f)
 	cfg.Cache.Redis.RegisterFlagsWithPrefix(prefix+"cache.redis.", f)
 }
@@ -58,6 +61,15 @@ func (cfg *Config) IsDefaults() bool {
 
 	// Note: cmp.Equal will panic if it encounters anything it cannot handle.
 	return cmp.Equal(*cfg, defaults, cmp.FilterPath(filterNonYaml, cmp.Ignore()), cmp.Comparer(equalSecrets))
+}
+
+type LocalStoreConfig struct {
+	Directory string `yaml:"directory"`
+}
+
+// RegisterFlagsWithPrefix registers flags with the input prefix.
+func (cfg *LocalStoreConfig) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet) {
+	f.StringVar(&cfg.Directory, prefix+"local.directory", "", "Directory to scan for rules")
 }
 
 // Return true if the path contains a struct field with tag `yaml:"-"`.