+++ pkg/ruler/service_discovery_test.go
@@ -68,14 +68,16 @@ func TestConfig_TranslatesToPrometheusTargetGroup(t *testing.T) {
 			resolver := &mockResolver{}
 			resolver.expectAnyResolveCall()
 			resolver.returnAddresses(tc.resolvedAddresses)
-
+			reg := prometheus.NewRegistry()
+			refreshMetrics := discovery.NewRefreshMetrics(reg)
 			cfg := dnsServiceDiscovery{
 				RefreshInterval: time.Millisecond,
 				Resolver:        resolver,
 				QType:           dns.A,
 				Host:            sourceAddress,
+				refreshMetrics:  refreshMetrics,
 			}
-			discoverer, err := cfg.NewDiscoverer(discovery.DiscovererOptions{Registerer: prometheus.NewPedanticRegistry()})
+			discoverer, err := cfg.NewDiscoverer(discovery.DiscovererOptions{})
 			require.NoError(t, err)
 
 			ctx, cancel := context.WithCancel(context.Background())
@@ -124,13 +126,16 @@ func TestConfig_ConstructsLookupNamesCorrectly(t *testing.T) {
 			resolver.expectResolveCalledWith(tc.expectedAddress)
 			resolver.returnAddresses(nil)
 
+			reg := prometheus.NewRegistry()
+			refreshMetrics := discovery.NewRefreshMetrics(reg)
 			cfg := dnsServiceDiscovery{
 				RefreshInterval: time.Millisecond,
 				Resolver:        resolver,
 				QType:           tc.qType,
 				Host:            tc.host,
+				refreshMetrics:  refreshMetrics,
 			}
-			discoverer, err := cfg.NewDiscoverer(discovery.DiscovererOptions{Registerer: prometheus.NewPedanticRegistry()})
+			discoverer, err := cfg.NewDiscoverer(discovery.DiscovererOptions{})
 			require.NoError(t, err)
 
 			ctx, cancel := context.WithCancel(context.Background())